requirements.txt
Flask
requests
python-dotenv
Flask-SQLAlchemy
mysql-connector-python